Scheduler with parsing ready

+++ src/main/java/com/vvopaa/zinterstraf/repository/ClientDao.java
@@ -3,7 +3,7 @@ package com.vvopaa.zinterstraf.repository;
 import com.vvopaa.zinterstraf.model.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientDao extends JpaRepository<Client, Integer> {
+public interface ClientDao extends JpaRepository<Client, Long> {
 
     Client findByClientId(String clientId);
 }
