+++ src/main/java/com/vvopaa/zinterstraf/repository/ProPlayerDao.java
@@ -3,6 +3,6 @@ package com.vvopaa.zinterstraf.repository;
 import com.vvopaa.zinterstraf.model.ProPlayer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProPlayerDao extends JpaRepository<ProPlayer, Integer> {
+public interface ProPlayerDao extends JpaRepository<ProPlayer, Long> {
 
 }
