+++ src/main/java/com/vvopaa/zinterstraf/repository/ProTeamDao.java
@@ -3,6 +3,6 @@ package com.vvopaa.zinterstraf.repository;
 import com.vvopaa.zinterstraf.model.ProTeam;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProTeamDao extends JpaRepository<ProTeam, Integer> {
+public interface ProTeamDao extends JpaRepository<ProTeam, Long> {
 
 }
