+++ src/main/java/com/vvopaa/zinterstraf/repository/UserRoleDao.java
@@ -1,12 +1,10 @@
 package com.vvopaa.zinterstraf.repository;
 
-import java.util.List;
 
 import com.vvopaa.zinterstraf.model.UserRoles;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRoleDao extends JpaRepository<UserRoles, Integer> {
+public interface UserRoleDao extends JpaRepository<UserRoles, Long> {
 
     UserRoles findByRole(String role);
-
 }
